New 2.5.0-2 toolchain does not like uncast bitwise

+++ lib/esp-knx-ip-0.5.1/esp-knx-ip-send.cpp
@@ -133,7 +133,7 @@ void ESPKNXIP::send_2byte_float(address_t const &receiver, knx_command_type_t ct
 	++e;
 	for (; v > 2047.0f; v /= 2)
 	++e;
-	long m = round(v) & 0x7FF;
+	long m = (long)round(v) & 0x7FF;
 	short msb = (short) (e << 3 | m >> 8);
 	if (val < 0.0f)
 	msb |= 0x80;
